+++ src/ForTests/Models/Core/CarCore.php
@@ -7,8 +7,8 @@ use Przeslijmi\Shortquery\Exceptions\Data\CollectionSliceNotPossibleException;
 use Przeslijmi\Shortquery\ForTests\Models\Car;
 use Przeslijmi\Shortquery\ForTests\Models\Cars;
 use Przeslijmi\Shortquery\ForTests\Models\Core\CarModel;
-use Przeslijmi\Shortquery\ForTests\Models\Girl as Girl3648;
-use Przeslijmi\Shortquery\ForTests\Models\Girls as Girls6973;
+use Przeslijmi\Shortquery\ForTests\Models\Girl as Girl9510;
+use Przeslijmi\Shortquery\ForTests\Models\Girls as Girls1305;
 use Przeslijmi\Shortquery\Tools\InstancesFactory;
 use stdClass;
 
@@ -426,24 +426,25 @@ class CarCore extends Instance
     /**
      * Returns child-Instance (one and only - for hasOne Relation type) in Relation.
      *
-     * @return Girl3648
+     * @return Girl9510
      */
-    public function getOneOwnerGirl() : Girl3648
+    public function getOneOwnerGirl() : Girl9510
     {
 
         return $this->oneOwnerGirl->getOne();
     }
 
     /**
-     * Call to add children (Girl3648) to this Instance.
+     * Call to add children (Girl9510) to this Instance.
      *
+     * @since  v1.0
      * @return self
      */
     public function expandOneOwnerGirl() : self
     {
 
         // Get records with those pks.
-        $child = new Girl3648(...func_get_args());
+        $child = new Girl9510(...func_get_args());
 
         // If we know that we need this one - read this one.
         if ($this->getOwnerGirl() !== null) {
@@ -460,16 +461,16 @@ class CarCore extends Instance
     /**
      * Adds one child-Instance to Relation Collection.
      *
-     * @param Girl3648 $instance One child-Instance of child for Relation.
+     * @param Girl9510 $instance One child-Instance of child for Relation.
      *
      * @return self
      */
-    public function addOneOwnerGirl(Girl3648 $instance) : self
+    public function addOneOwnerGirl(Girl9510 $instance) : self
     {
 
         // If there is no Collection created - create one.
         if (is_null($this->oneOwnerGirl) === true) {
-            $this->oneOwnerGirl = new Girls6973();
+            $this->oneOwnerGirl = new Girls1305();
         }
 
         // Put this Instance to this Collection.
@@ -481,24 +482,25 @@ class CarCore extends Instance
     /**
      * Returns child-Instance (one and only - for hasOne Relation type) in Relation.
      *
-     * @return Girl3648
+     * @return Girl9510
      */
-    public function getOneOwnerGirlWithSnapchat() : Girl3648
+    public function getOneOwnerGirlWithSnapchat() : Girl9510
     {
 
         return $this->oneOwnerGirlWithSnapchat->getOne();
     }
 
     /**
-     * Call to add children (Girl3648) to this Instance.
+     * Call to add children (Girl9510) to this Instance.
      *
+     * @since  v1.0
      * @return self
      */
     public function expandOneOwnerGirlWithSnapchat() : self
     {
 
         // Get records with those pks.
-        $child = new Girl3648(...func_get_args());
+        $child = new Girl9510(...func_get_args());
 
         // If we know that we need this one - read this one.
         if ($this->getOwnerGirl() !== null) {
@@ -515,16 +517,16 @@ class CarCore extends Instance
     /**
      * Adds one child-Instance to Relation Collection.
      *
-     * @param Girl3648 $instance One child-Instance of child for Relation.
+     * @param Girl9510 $instance One child-Instance of child for Relation.
      *
      * @return self
      */
-    public function addOneOwnerGirlWithSnapchat(Girl3648 $instance) : self
+    public function addOneOwnerGirlWithSnapchat(Girl9510 $instance) : self
     {
 
         // If there is no Collection created - create one.
         if (is_null($this->oneOwnerGirlWithSnapchat) === true) {
-            $this->oneOwnerGirlWithSnapchat = new Girls6973();
+            $this->oneOwnerGirlWithSnapchat = new Girls1305();
         }
 
         // Put this Instance to this Collection.